* commit '0c85d63ef65081908a2e475ddeac837d181acbfd':
  Victor added posibility for Company to add service,invoice. Display all his services

+++ src/main/java/com/endava/service_system/controller/rest/ServiceRestController.java
@@ -55,11 +55,11 @@ public class ServiceRestController {
 //    }
 
     @GetMapping("/services")
-    public List<ServiceToUserDto> getServices(@RequestParam(value = "categoryId", required = false) Integer categoryId,
+    public List<ServiceToUserDto> getServices(@RequestParam(value = "categoryId", required = false) Long categoryId,
                                               @RequestParam(value = "size", required = false) Integer size,
                                               @RequestParam(value = "page", required = false) Integer page,
                                               @RequestParam(required = false, value = "min") Integer min,
-                                              @RequestParam(required = false, value = "companyId") Integer companyId,
+                                              @RequestParam(required = false, value = "companyId") Long companyId,
                                               @RequestParam(required = false, value = "company") String companyName,
                                               @RequestParam(required = false, value = "category") String categoryName,
                                               @RequestParam(required = false, value = "max") Integer max,
